CMake: Unconditional unit test support
